+++ __tests__/require-usememo-children.ts
@@ -1,5 +1,5 @@
 import { RuleTester } from "eslint";
-import rule from "../require-usememo-children";
+import rule from "src/require-usememo-children";
 
 const ruleTester = new RuleTester({
   parser: require.resolve("@typescript-eslint/parser"),
