+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
   "compilerOptions": {
     "strict": true,
-    "outDir": "dist"
+    "outDir": "dist",
+    "moduleResolution": "node",
+    "baseUrl": "./",
   },
-  "exclude": ["__tests__"]
 }
